Update comparison with Mesa
